requirements.txt
transformers
sqlalchemy
ijson
kss
namuwiki-extractor

